+++ go.mod
@@ -2,6 +2,9 @@ module github.com/essentialkaos/deadline
 
 go 1.18
 
-require github.com/essentialkaos/ek/v12 v12.93.0
+require (
+	github.com/essentialkaos/depsy v1.1.0
+	github.com/essentialkaos/ek/v12 v12.93.0
+)
 
 require golang.org/x/sys v0.16.0 // indirect
